helped refactor some components
