+++ screens/HomeScreen.jsx
@@ -3,23 +3,18 @@ import React from "react";
 import { useNavigation } from "@react-navigation/native";
 import { signOut } from "firebase/auth";
 import { auth } from "../firebase/firebase";
+import { handleSignOut } from "../firebase/auth/signout";
 
 const HomeScreen = () => {
   const { replace } = useNavigation();
-  const handleSignOut = () => {
-    signOut(auth)
-      .then(() => {
-        alert("signed out successful!");
-        replace("Login");
-      })
-      .catch((err) => {
-        alert(`${err.message}`);
-      });
-  };
+
   return (
     <View style={styles.container}>
       <Text>Email: {auth.currentUser?.email}</Text>
-      <TouchableOpacity style={styles.button} onPress={handleSignOut}>
+      <TouchableOpacity
+        style={styles.button}
+        onPress={() => handleSignOut(replace)}
+      >
         <Text style={styles.buttonText}>Sign Out</Text>
       </TouchableOpacity>
     </View>
